tests/requirements.txt
requests
deepdiff
mcp==1.12.2
uvicorn[standard]>=0.30.0
pyyaml
pydantic>=2.0.0
httpx
google-genai